+++ core/customizer/config.php
@@ -224,7 +224,14 @@ class Layers_Customizer_Config {
 									'placeholder'	=> ".classname {\n\tbackground: #333;\n}",
 									'sanitize_callback' => FALSE
 								),
+								'upsell-devkit-heading' => array(
+									'type'  => 'layers-heading',
+									'class' => 'layers-upsell-tag',
+									'label'    => __( 'Upgrade to DevKit' , 'layerswp' ),
+									'description' => __( 'Want the best CSS customization interface? <a target="_blank" href="http://codecanyon.net/item/devkit-developer-tools-for-layers-/11221313?ref=obox">Purchase DevKit</a> and save bundles of time!' , 'layerswp' ),
+								),
 							); // css
+
 
 		// Header -> Layout -> Layout
 		$controls['fonts'] = array(
@@ -448,9 +455,9 @@ class Layers_Customizer_Config {
 					),
 					'upsell-colorkit-heading' => array(
 						'type'  => 'layers-heading',
-						'class' => 'layers-usell-tag',
+						'class' => 'layers-upsell-tag',
 						'label'    => __( 'Upgrade to ColorKit' , 'layerswp' ),
-						'description' => __( 'Want more color customzation? <a target="_blank" href="http://codecanyon.net/item/colorkit-color-customization-for-layers/11225042">Purchase ColorKit</a> and get the full box of crayons!' , 'layerswp' ),
+						'description' => __( 'Want more color customzation? <a target="_blank" href="http://codecanyon.net/item/colorkit-color-customization-for-layers/11225042?ref=obox">Purchase ColorKit</a> and get the full box of crayons!' , 'layerswp' ),
 					),
 				);